remove property `normalised_peaks`
